+++ conformity/opencv/test_cvtcolor.h
@@ -1,0 +1,14 @@
+// SPDX-FileCopyrightText: 2024 Arm Limited and/or its affiliates <open-source-office@arm.com>
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#ifndef KLEIDICV_OPENCV_CONFORMITY_TEST_CVTCOLOR_H_
+#define KLEIDICV_OPENCV_CONFORMITY_TEST_CVTCOLOR_H_
+
+#include <vector>
+
+#include "tests.h"
+
+std::vector<test>& cvtcolor_tests_get();
+
+#endif  // KLEIDICV_OPENCV_CONFORMITY_TEST_CVTCOLOR_H_